\name{orthgenes}
\docType{data}
\alias{orthgenes}
\title{A real dataset of orthologous genes between the different species.}
\description{
  This data set gives 27821 orthologous genes which include read counts and
  genes length between the two different species.
}
\usage{orthgenes}

\format{A data.frame containing 27821 orthologous genes.}
\source{
Brawand D, Soumillon M, Necsulea A, Julien P, Csardi G, Harrigan P, et al.
The evolution of gene expression levels in mammalian organs.
Nature. 2011;478:343-348.
}
\keyword{datasets}

