\name{sim_data}
\docType{data}
\alias{sim_data}
\title{A simulation dataset of orthologous genes between the different species.}
\description{
  This data set gives 4149 orthologous genes which include read counts and
  genes length between the two different species.
}
\usage{sim_data}

\format{A data.frame containing 4149 orthologous genes.}
\source{
  Zhou Y, Zhu JD, Tong TJ, Wang JH, Lin BQ, Zhang J(2018, pending publication).
A Novel Normalization Method and Differential Expression Analysis of RNA-seq
Data between Different Species.
}
\keyword{datasets}

